BUG/MAJOR: xodtemplate: Fix service dependency inheritance
This patch is for Naemon 1.0.3.

This bug was fixed in Naemon 1.0.5 with the switch from rbtree
to glib2 GTree. The comparison function is for servicedependency
templates, hence should compare the "name" property, not the
"host_name" and "service_description" properties. Other
xod_sort_compare_*_template() functions are implemented in that way.

I've added some debug before the patch. The entry for the "check_mk2"
servicedependency template gets inserted in the tree (0x2029e60). When
processing the configuration, Naemon hits a servicedependency (0x202ba90)
that uses the "check_mk2" template. It looks up the "check_mk2" entry
with xodtemplate_find_servicedependency() but the later does not find it
and returns NULL, leading to the following error message:

Error: Template 'check_mk2' specified in service dependency definition
could not be found (config file
'/home/admin/.pomgen/nagios-input/objects.cfg', starting on line 153)

>> rbtree_insert OBJTYPE_SERVICEDEPENDENCY 0x2029e60 name="check_mk2"
>> xodtemplate_resolve_servicedependency 0x202ba90 uses="check_mk2"
>> xodtemplate_find_servicedependency "check_mk2" = (nil)

After applying the patch, we get the expected result and the error is
gone:

>> rbtree_insert OBJTYPE_SERVICEDEPENDENCY 0x2591e60 name="check_mk2"
>> xodtemplate_resolve_servicedependency 0x2593a90 uses="check_mk2"
>> xodtemplate_find_servicedependency "check_mk2" = 0x2591e60