BUG/MAJOR: commands: Fix long_plugin_output support for passive checks
Passive checks with long_plugin_output and send throught the old style
single-line based interface are broken because special sequences "\\"
and "\n" are not unescaped anymore.

The issue applies to commands sent via the FIFO or read from a file as
a result of the PROCESS_FILE command. It does not applies to commands
sent via the Naemon query handler socket with the COMMAND_SYNTAX_NOKV
syntax because these are NULL-byte terminated, not single-line based.

The main goal for this commit is to maintain compatibility with Nagios
behavior, to ease migrations to Naemon.