BUG/MAJOR: checks: Restore long_plugin_output escape Nagios compatibility
In Naemon code, the commit 56098925d "checks: Refactor check output
parsing" breaks long_plugin_output behavior. Newline characters are not
escaped anymore, resulting in:
- potential broken status.dat parsing;
- potential broken behavior for tool querying long_plugin_output
  through livestatus.

The goal of this patch is to maintain compatibility with Nagios
behavior, to ease migrations to Naemon. With this patch, the
long_plugin_output property remains stored as a single line string
with backslahes escaped as "\\" (0x5C 0x5C) and with newlines
escaped as "\n" (0x5C 0x6E).