BUG/MEDIUM: checks: Missing update of service latency on passive checks
This patch is a workaround to preserve Nagios compatibility, certainly not
what Naemon developpers would like.

Check latency is set on the service in run_scheduled_service_check, so
only for active checks but we want a value on passive checks as well with
the same meaning as with Nagios. The latency is available in the
<queued_check_result> data, re-update it... but we know that something is
not right here...